requirements.txt
# Core dependencies
fastapi==0.95.2  # Version required by chromadb
uvicorn[standard]==0.22.0  # Compatible with fastapi 0.95.2
python-dotenv==1.0.0
pydantic==1.10.12  # Compatible with fastapi 0.95.2
PyYAML==6.0  # Using a newer version with pre-built wheels
python-multipart==0.0.6
aiofiles==23.1.0

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27

# Web crawling and processing
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8  # Keep this first as it's more flexible with lxml
trafilatura==1.6.2  # This requires lxml 4.9.2 on macOS
lxml==4.9.2  # Specific version for trafilatura compatibility
python-slugify==8.0.1
urllib3==1.26.16

# LLM and Embedding Providers
openai==0.27.8
huggingface-hub==0.14.1
transformers==4.28.1
sentence-transformers==2.2.2
safetensors==0.3.1
accelerate==0.18.0
bitsandbytes==0.39.1
ctransformers==0.2.27
llama-cpp-python==0.1.78

# Vector Stores and RAG
langchain==0.0.267  # Updated to a version compatible with pydantic v1
chromadb==0.4.15  # Updated to a version that works with newer FastAPI
faiss-cpu==1.7.4

# PyTorch (CPU version for compatibility)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Additional dependencies
tokenizers==0.13.3
tqdm==4.65.0

# Development dependencies
pytest==7.3.1
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.3.0
